Ensure a channel can be deleted multiple times (#363)
* Ensure a channel can be deleted multiple times

If the channel get deleted because a failure happened
on the remote, and than a similar failure happens on
the local as a result, the channel delete could happen
twice which would try to send the close message again
but the output queue is already destroyed after it was
sent the first time. This should not be a failure as
we want to be able to delete the channel any time there
is a failure.

* preen

* preen

* preen

* add missing coverage

* Update snitun/multiplexer/core.py

* Update snitun/multiplexer/core.py